+++ cmd/rr/main.go
@@ -23,9 +23,7 @@
 package main
 
 import (
-	// colorful logging
 	"github.com/sirupsen/logrus"
-	
 	rr "github.com/spiral/roadrunner/cmd/rr/cmd"
 
 	// services (plugins)
@@ -34,18 +32,17 @@ import (
 	"github.com/spiral/roadrunner/service/rpc"
 	"github.com/spiral/roadrunner/service/static"
 
-	// additional command handlers
+	// additional command and debug handlers
 	_ "github.com/spiral/roadrunner/cmd/rr/http"
 )
 
 func main() {
-	rr.Logger.Formatter = &logrus.TextFormatter{ForceColors: true}
-
-	rr.Container.Register(env.ID, env.NewService(map[string]string{"rr": rr.Version}))
-
+	rr.Container.Register(env.ID, &env.Service{})
 	rr.Container.Register(rpc.ID, &rpc.Service{})
 	rr.Container.Register(http.ID, &http.Service{})
 	rr.Container.Register(static.ID, &static.Service{})
+
+	rr.Logger.Formatter = &logrus.TextFormatter{ForceColors: true}
 
 	// you can register additional commands using cmd.CLI
 	rr.Execute()