+++ service/env/environment.go
@@ -2,7 +2,10 @@ package env
 
 // Provider aggregates list of environment variables. This interface can be used in custom implementation to drive
 // values from external sources.
-type Provider interface {
+type Environment interface {
 	// GetEnv must return list of env variables.
 	GetEnv() (map[string]string, error)
+
+	// SetEnv sets or creates environment value.
+	SetEnv(key, value string)
 }