+++ service/http/rpc.go
@@ -29,7 +29,7 @@ type Worker struct {
 
 // Reset resets underlying RR worker pool and restarts all of it's workers.
 func (rpc *rpcServer) Reset(reset bool, r *string) error {
-	if rpc.svc.srv == nil {
+	if rpc.svc == nil || rpc.svc.srv == nil {
 		return errors.New("http server is not running")
 	}
 
@@ -39,7 +39,7 @@ func (rpc *rpcServer) Reset(reset bool, r *string) error {
 
 // Workers returns list of active workers and their stats.
 func (rpc *rpcServer) Workers(list bool, r *WorkerList) error {
-	if rpc.svc.srv == nil {
+	if rpc.svc == nil || rpc.svc.srv == nil {
 		return errors.New("http server is not running")
 	}